Compile CLI to wasm
