+++ Cargo.toml
@@ -25,25 +25,25 @@ path = "src/bin/wasmtime.rs"
 doc = false
 
 [dependencies]
-wasmtime = { workspace = true }
+wasmtime = { workspace = true, default-features = false, features = [] }
 wasmtime-cache = { workspace = true, optional = true }
 wasmtime-cli-flags = { workspace = true }
 wasmtime-cranelift = { workspace = true, optional = true }
 wasmtime-environ = { workspace = true }
 wasmtime-explorer = { workspace = true, optional = true }
 wasmtime-wast = { workspace = true, optional = true }
-wasmtime-wasi = { workspace = true, default-features = true, features = [
+wasmtime-wasi = { workspace = true, optional = true, default-features = true, features = [
   "exit",
 ] }
 wasmtime-wasi-nn = { workspace = true, optional = true }
 wasmtime-wasi-threads = { workspace = true, optional = true }
 wasmtime-wasi-http = { workspace = true, optional = true }
-wasmtime-runtime = { workspace = true }
+wasmtime-runtime = { workspace = true, optional = true }
 clap = { workspace = true }
 anyhow = { workspace = true }
 target-lexicon = { workspace = true }
 once_cell = { workspace = true }
-listenfd = "1.0.0"
+listenfd = { version = "1.0.0", optional = true }
 wat = { workspace = true, optional = true }
 serde = { workspace = true }
 serde_derive = { workspace = true }
@@ -306,6 +306,7 @@ default = [
   # All subcommands are included by default.
   "compile",
   "explore",
+  "run",
   "serve",
   "wast",
   "config",
@@ -328,6 +329,7 @@ default = [
   "addr2line",
   "debug-builtins",
   "component-model",
+  "async",
 
   # Enable some nice features of clap by default, but they come at a binary size
   # cost, so allow disabling this through disabling of our own `default`
@@ -375,17 +377,19 @@ profiling = ["wasmtime/profiling"]
 coredump = ["wasmtime-cli-flags/coredump"]
 addr2line = ["wasmtime/addr2line"]
 debug-builtins = ["wasmtime/debug-builtins"]
+async = ["wasmtime/async"]
 
 # Enables compatibility shims with Wasmtime 13 and prior's CLI.
 old-cli = []
 
 # CLI subcommands for the `wasmtime` executable. See `wasmtime $cmd --help`
 # for more information on each subcommand.
-serve = ["wasi-http", "component-model", "dep:http-body-util", "dep:http"]
+serve = ["wasi-http", "component-model", "dep:http-body-util", "dep:http", "run"]
 explore = ["dep:wasmtime-explorer"]
 wast = ["dep:wasmtime-wast"]
 config = ["cache"]
 compile = ["cranelift"]
+run = ["wasmtime/runtime", "dep:wasmtime-wasi", "dep:listenfd", "dep:wasmtime-runtime"]
 
 [[test]]
 name = "host_segfault"
