+++ run.sh
@@ -1,0 +1,4 @@
+❯ cargo +nightly --version
+cargo 1.78.0-nightly (ccc84ccec 2024-02-07)
+
+❯ cargo +nightly build -p wasmtime-cli --no-default-features --target wasm32-wasi --features all-arch
