+++ src/commands/serve.rs
@@ -1,4 +1,4 @@
-use crate::common::{Profile, RunCommon, RunTarget};
+use crate::commands::run::{Profile, RunCommon, RunTarget};
 use anyhow::{anyhow, bail, Result};
 use clap::Parser;
 use std::net::SocketAddr;
